// frontend/types/dataTypeEnum.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * My API
 * Documentação automática da API
 * OpenAPI spec version: 1.0.0
 */

/**
 * * `quantitative` - Quantitative
* `categorical` - Categorical
 */
export type DataTypeEnum = typeof DataTypeEnum[keyof typeof DataTypeEnum];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const DataTypeEnum = {
  quantitative: 'quantitative',
  categorical: 'categorical',
} as const;